refactor: split control plane router modules
