+++ packages/control-plane/src/routes/provider-identities.ts
@@ -17,10 +17,6 @@ type UpsertProviderIdentityRequest = {
   avatarUrl?: unknown;
 };
 
-function isObjectRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
 function optionalString(value: unknown): string | undefined {
   if (typeof value !== "string") return undefined;
   const trimmed = value.trim();
@@ -44,8 +40,10 @@ export async function handleUpsertProviderIdentity(
 ): Promise<Response> {
   const body = await parseJsonBody<UpsertProviderIdentityRequest>(request);
   if (body instanceof Response) return body;
-  if (!isObjectRecord(body)) {
-    return error("Request body must be an object", 400);
+
+  const provider = pathSegment(match.groups?.provider);
+  if (provider !== "github") {
+    return error("provider must be github", 400);
   }
 
   const providerUserId = pathSegment(match.groups?.providerUserId);
@@ -73,7 +71,7 @@ export async function handleUpsertProviderIdentity(
 export const providerIdentityRoutes: Route[] = [
   {
     method: "PUT",
-    pattern: parsePattern("/provider-identities/github/:providerUserId"),
+    pattern: parsePattern("/provider-identities/:provider/:providerUserId"),
     handler: handleUpsertProviderIdentity,
   },
 ];
