+++ packages/control-plane/src/routes/session-index.ts
@@ -1,0 +1,94 @@
+import { isCanonicalUserId, type SessionStatus } from "@open-inspect/shared";
+import { SessionIndexStore } from "../db/session-index";
+import { error, json, parsePattern, type RequestContext, type Route } from "./shared";
+import type { Env } from "../types";
+
+const SESSION_STATUSES: SessionStatus[] = [
+  "created",
+  "active",
+  "completed",
+  "failed",
+  "archived",
+  "cancelled",
+];
+
+function parseSessionStatus(value: string | null): SessionStatus | undefined {
+  if (!value) return undefined;
+  return SESSION_STATUSES.includes(value as SessionStatus) ? (value as SessionStatus) : undefined;
+}
+
+function parseCreatedByFilters(searchParams: URLSearchParams): string[] | Response {
+  const values = searchParams.getAll("createdBy");
+  const userIds: string[] = [];
+  const seen = new Set<string>();
+
+  for (const value of values) {
+    if (!isCanonicalUserId(value)) {
+      return error("Invalid createdBy", 400);
+    }
+
+    if (!seen.has(value)) {
+      seen.add(value);
+      userIds.push(value);
+    }
+  }
+
+  return userIds;
+}
+
+async function handleListSessions(
+  request: Request,
+  env: Env,
+  _match: RegExpMatchArray,
+  _ctx: RequestContext
+): Promise<Response> {
+  const url = new URL(request.url);
+  const limit = Math.min(parseInt(url.searchParams.get("limit") || "50"), 100);
+  const offset = parseInt(url.searchParams.get("offset") || "0");
+  const statusParam = url.searchParams.get("status");
+  const excludeStatusParam = url.searchParams.get("excludeStatus");
+  const status = parseSessionStatus(statusParam);
+  const excludeStatus = parseSessionStatus(excludeStatusParam);
+  const createdByUserIds = parseCreatedByFilters(url.searchParams);
+
+  if (statusParam && !status) {
+    return error("Invalid status", 400);
+  }
+
+  if (excludeStatusParam && !excludeStatus) {
+    return error("Invalid excludeStatus", 400);
+  }
+
+  if (createdByUserIds instanceof Response) {
+    return createdByUserIds;
+  }
+
+  const store = new SessionIndexStore(env.DB);
+  const result = await store.list({ status, excludeStatus, createdByUserIds, limit, offset });
+
+  return json({
+    sessions: result.sessions,
+    total: result.total,
+    hasMore: result.hasMore,
+  });
+}
+
+async function handleDeleteSession(
+  _request: Request,
+  env: Env,
+  match: RegExpMatchArray,
+  _ctx: RequestContext
+): Promise<Response> {
+  const sessionId = match.groups?.id;
+  if (!sessionId) return error("Session ID required");
+
+  const sessionStore = new SessionIndexStore(env.DB);
+  await sessionStore.delete(sessionId);
+
+  return json({ status: "deleted", sessionId });
+}
+
+export const sessionIndexRoutes: Route[] = [
+  { method: "GET", pattern: parsePattern("/sessions"), handler: handleListSessions },
+  { method: "DELETE", pattern: parsePattern("/sessions/:id"), handler: handleDeleteSession },
+];
