+++ packages/control-plane/src/routes/session-media.ts
@@ -1,0 +1,8 @@
+import type { Route } from "./shared";
+import { sessionMediaStreamRoutes } from "./session-media-stream";
+import { sessionMediaUploadRoutes } from "./session-media-upload";
+
+export const sessionMediaRoutes: Route[] = [
+  ...sessionMediaUploadRoutes,
+  ...sessionMediaStreamRoutes,
+];
