+++ packages/control-plane/src/routes/session-prompt.ts
@@ -1,0 +1,97 @@
+import type { CallbackContext } from "@open-inspect/shared";
+import { SessionIndexStore } from "../db/session-index";
+import { UserStore } from "../db/user-store";
+import { createLogger } from "../logger";
+import { SessionInternalPaths } from "../session/contracts";
+import { parseAuthorId, resolveGitHubEnrichment, type GitHubEnrichment } from "../session/identity";
+import type { Env } from "../types";
+import { error, parsePattern, type Route } from "./shared";
+import { sessionRoute, type SessionRouteContext } from "./session-route";
+
+const logger = createLogger("router:session-prompt");
+
+async function handleSessionPrompt(
+  request: Request,
+  env: Env,
+  match: RegExpMatchArray,
+  ctx: SessionRouteContext
+): Promise<Response> {
+  const sessionId = match.groups?.id;
+  if (!sessionId) return error("Session ID required");
+
+  const body = (await request.json()) as {
+    content: string;
+    authorId?: string;
+    source?: string;
+    model?: string;
+    reasoningEffort?: string;
+    attachments?: Array<{ type: string; name: string; url?: string }>;
+    callbackContext?: CallbackContext;
+  };
+
+  if (!body.content) {
+    return error("content is required");
+  }
+
+  const authorId = body.authorId || "anonymous";
+
+  let enrichment: GitHubEnrichment | undefined;
+  const parsed = parseAuthorId(authorId);
+  if (parsed) {
+    try {
+      const userStore = new UserStore(env.DB);
+      const identity = await userStore.getIdentity(parsed.provider, parsed.providerUserId);
+      if (identity) {
+        enrichment = (await resolveGitHubEnrichment(env, userStore, identity.userId)) ?? undefined;
+      }
+    } catch (e) {
+      logger.warn("Failed to enrich prompt with GitHub identity", {
+        error: e instanceof Error ? e : String(e),
+        authorId,
+      });
+    }
+  }
+
+  const response = await ctx.sessionRuntime.fetch(sessionId, SessionInternalPaths.prompt, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      content: body.content,
+      authorId,
+      source: body.source || "web",
+      model: body.model,
+      reasoningEffort: body.reasoningEffort,
+      attachments: body.attachments,
+      callbackContext: body.callbackContext,
+      authorDisplayName: enrichment?.displayName,
+      authorEmail: enrichment?.email,
+      authorLogin: enrichment?.scmLogin,
+      scmUserId: enrichment?.scmUserId,
+      scmAccessTokenEncrypted: enrichment?.accessTokenEncrypted,
+      scmRefreshTokenEncrypted: enrichment?.refreshTokenEncrypted,
+      scmTokenExpiresAt: enrichment?.tokenExpiresAt,
+    }),
+  });
+
+  const store = new SessionIndexStore(env.DB);
+  ctx.executionCtx?.waitUntil(
+    store.touchUpdatedAt(sessionId).catch((error) => {
+      logger.error("session_index.touch_updated_at.background_error", {
+        session_id: sessionId,
+        trace_id: ctx.trace_id,
+        request_id: ctx.request_id,
+        error,
+      });
+    })
+  );
+
+  return response;
+}
+
+export const sessionPromptRoutes: Route[] = [
+  sessionRoute({
+    method: "POST",
+    pattern: parsePattern("/sessions/:id/prompt"),
+    handler: handleSessionPrompt,
+  }),
+];
