+++ packages/control-plane/src/session/contracts.test.ts
@@ -5,38 +5,30 @@ import { SessionInternalPaths } from "./contracts";
 describe("session internal endpoint contracts", () => {
   it("uses contract constants in internal route wiring and router for known endpoints", () => {
     const routerSource = readFileSync(new URL("../router.ts", import.meta.url), "utf8");
+    const sessionRouteSources = [
+      "../routes/session-create.ts",
+      "../routes/session-runtime-proxy.ts",
+      "../routes/session-ws-token.ts",
+      "../routes/session-prompt.ts",
+      "../routes/session-media.ts",
+      "../routes/session-media-artifacts.ts",
+      "../routes/session-media-stream.ts",
+      "../routes/session-media-upload.ts",
+      "../routes/session-child-spawn.ts",
+      "../routes/session-children.ts",
+    ]
+      .map((path) => readFileSync(new URL(path, import.meta.url), "utf8"))
+      .join("\n");
+    const runtimeClientSource = readFileSync(
+      new URL("./runtime-client.ts", import.meta.url),
+      "utf8"
+    );
     const initializeSource = readFileSync(new URL("./initialize.ts", import.meta.url), "utf8");
     const routesSource = readFileSync(new URL("./http/routes.ts", import.meta.url), "utf8");
     const durableObjectSource = readFileSync(
       new URL("./durable-object.ts", import.meta.url),
       "utf8"
     );
-
-    // Endpoints used directly in router.ts
-    const routerEndpointKeys: Array<keyof typeof SessionInternalPaths> = [
-      "verifySandboxToken",
-      "state",
-      "prompt",
-      "stop",
-      "createMediaArtifact",
-      "events",
-      "artifacts",
-      "participants",
-      "messages",
-      "createPr",
-      "openaiTokenRefresh",
-      "wsToken",
-      "archive",
-      "unarchive",
-      "spawnContext",
-      "childSessionUpdate",
-      "childSummary",
-      "cancel",
-    ];
-
-    for (const endpointKey of routerEndpointKeys) {
-      expect(routerSource).toContain(`SessionInternalPaths.${endpointKey}`);
-    }
 
     // "init" is used in session/initialize.ts (extracted from router)
     expect(initializeSource).toContain("SessionInternalPaths.init");
@@ -48,7 +40,10 @@ describe("session internal endpoint contracts", () => {
     }
 
     expect(durableObjectSource).toContain("createSessionInternalRoutes");
-    expect(routerSource).not.toContain("http://internal/internal/");
+    expect(runtimeClientSource).toContain("buildSessionInternalUrl");
+    const externalSessionSource = `${routerSource}\n${sessionRouteSources}`;
+    expect(externalSessionSource).not.toContain("http://internal/internal/");
+    expect(runtimeClientSource).not.toContain("http://internal/internal/");
     expect(routesSource).not.toContain('"/internal/');
     expect(routesSource).not.toContain("'/internal/");
   });
